+++ lib/src/state_management.dart
@@ -4,3 +4,4 @@ export 'state_management/slim_app_state_controller.dart';
 export 'state_management/slim_builder.dart';
 export 'state_management/slim_controller.dart';
 export 'state_management/slimer.dart';
+export 'state_management/slim_widget.dart';